Add files via upload
To Run Install all the necessary libraries above and then just execute [43]" 
! streamlit run newapp.py & npx localtunnel --port 8501"

Rerun the streamlit app and it will start working. There are two ways to input. A dropdown at the left and text input field in the center.

A Restaurant name and probable menu items appear at the end

Enjoy!!